there should only be one DateUtil class
