+++ app/src/main/java/it/niedermann/nextcloud/deck/util/DateUtil.java
@@ -4,6 +4,7 @@ import android.content.Context;
 import android.text.format.DateUtils;
 
 import java.util.Date;
+import java.util.TimeZone;
 import java.util.concurrent.TimeUnit;
 
 import it.niedermann.nextcloud.deck.R;
@@ -12,6 +13,27 @@ public final class DateUtil {
     private static final int DATE_TIME_PARTS_SIZE = 2;
 
     private DateUtil() {
+    }
+
+    public static Date nowInGMT() {
+        return convertToGMT(new Date());
+    }
+
+    private static Date convertToGMT(Date date ){
+        TimeZone tz = TimeZone.getDefault();
+        Date ret = new Date( date.getTime() - tz.getRawOffset() );
+
+        // if we are now in DST, back off by the delta.  Note that we are checking the GMT date, this is the KEY.
+        if ( tz.inDaylightTime( ret )){
+            Date dstDate = new Date( ret.getTime() - tz.getDSTSavings() );
+
+            // check to make sure we have not crossed back into standard time
+            // this happens when we are on the cusp of DST (7pm the day before the change for PDT)
+            if ( tz.inDaylightTime( dstDate )){
+                ret = dstDate;
+            }
+        }
+        return ret;
     }
 
     /**
